Replace individual module calls with GENERATE_MUTATED_PEPTIDES subworkflow and fix FASTA headers
- Add new modules: generatemutfasta/1.2, mutalyzer/retriever, neosv
- Add GENERATE_MUTATED_PEPTIDES subworkflow combining mutalyzer cache
  generation, mutant FASTA creation, SV neoantigen prediction, and
  HLA string generation
- Remove old neoantigenutils/generatemutfasta module
- Fix FASTA headers to use identifier_key with _M/_W suffix (stays
  under netMHCpan 15-char Identity truncation limit)
- Update modules.json to point generatemutfasta/1.2 at develop branch
- Update workflow, configs, and test snapshots
