+++ assets/samplesheet.csv
@@ -1,3 +1,3 @@
 sample,maf,facets_hisens_cncf,hla_file
-tumor_normal,https://raw.githubusercontent.com/mskcc-omics-workflows/test-datasets/neoantigen/neoantigen/temp_test_somatic_unfiltered.maf,https://raw.githubusercontent.com/mskcc-omics-workflows/test-datasets/neoantigen/neoantigen/facets_hisens.cncf.txt,https://raw.githubusercontent.com/mskcc-omics-workflows/test-datasets/neoantigen/neoantigen/winners.hla.txt
-tumor_normal2,https://raw.githubusercontent.com/mskcc-omics-workflows/test-datasets/neoantigen/neoantigen/temp_test_somatic_unfiltered.maf,https://raw.githubusercontent.com/mskcc-omics-workflows/test-datasets/neoantigen/neoantigen/facets_hisens.cncf.txt,https://raw.githubusercontent.com/mskcc-omics-workflows/test-datasets/neoantigen/neoantigen/winners.hla.txt
+tumor_normal,https://raw.githubusercontent.com/mskcc-omics-workflows/test-datasets/neoantigen/neoantigen/small_test_chr2and16/somatic_filtered.chr2_16.maf,https://raw.githubusercontent.com/mskcc-omics-workflows/test-datasets/neoantigen/neoantigen/small_test_chr2and16/facets_hisens.cncf.chr2_16.txt,https://raw.githubusercontent.com/mskcc-omics-workflows/test-datasets/neoantigen/neoantigen/winners.hla.txt
+tumor_normal2,https://raw.githubusercontent.com/mskcc-omics-workflows/test-datasets/neoantigen/neoantigen/small_test_chr2and16/somatic_filtered.chr2_16.maf,https://raw.githubusercontent.com/mskcc-omics-workflows/test-datasets/neoantigen/neoantigen/small_test_chr2and16/facets_hisens.cncf.chr2_16.txt,https://raw.githubusercontent.com/mskcc-omics-workflows/test-datasets/neoantigen/neoantigen/winners.hla.txt
