+++ conf/modules.config
@@ -17,6 +17,10 @@ process {
         saveAs: { filename -> filename.equals('versions.yml') ? null : filename }
     ]
 
+    withName: 'MUTALYZER_RETRIEVER' {
+        storeDir = "${params.outdir}/mutalyzer_cache"
+    }
+
     withName: 'PHYLOWGS_PARSECNVS' {
         ext.args = '-f facets'
     }
