+++ conf/prod.config
@@ -43,7 +43,8 @@ params {
     compute_fitness_w = 0.22402192838740312
 
     iedbfasta = 'https://raw.githubusercontent.com/mskcc/NeoantigenEditing/refs/heads/main/data/iedb.fasta'
-    cds = 'https://github.com/mskcc-omics-workflows/test-datasets/raw/neoantigen/neoantigen/Homo_sapiens.GRCh37.75.cds.all.fa.gz'
+    reference_fasta = null // TODO: set URL/path to reference genome FASTA (bgzipped) for GRCh37
+    reference_gff3  = null // TODO: set URL/path to reference GFF3 annotation for GRCh37
     cdna = 'https://github.com/mskcc-omics-workflows/test-datasets/raw/neoantigen/neoantigen/Homo_sapiens.GRCh37.75.cdna.all.fa.gz'
     gtf = 'https://github.com/mskcc-omics-workflows/test-datasets/raw/neoantigen/neoantigen/Homo_sapiens.GRCh37.75.gtf.gz'
 }
