+++ conf/test.config
@@ -39,7 +39,8 @@ params {
     compute_fitness_w = 0.22402192838740312
 
     iedbfasta = 'https://raw.githubusercontent.com/mskcc-omics-workflows/test-datasets/neoantigen/neoantigen/neoantigenEditing/data/iedb.fasta'
-    cds = 'https://github.com/mskcc-omics-workflows/test-datasets/raw/neoantigen/neoantigen/Homo_sapiens.GRCh37.75.cds.all.fa.gz'
+    reference_fasta = 'https://github.com/mskcc-omics-workflows/test-datasets/raw/neoantigen/neoantigen/small_test_chr2and16/Homo_sapiens.GRCh37.dna_rm.chromosome.chr2_16.fa.bgzip.gz'
+    reference_gff3  = 'https://github.com/mskcc-omics-workflows/test-datasets/raw/neoantigen/neoantigen/small_test_chr2and16/Homo_sapiens.GRCh37.87.chr2_16.gff3'
     cdna = 'https://github.com/mskcc-omics-workflows/test-datasets/raw/neoantigen/neoantigen/Homo_sapiens.GRCh37.75.cdna.all.fa.gz'
     gtf = 'https://github.com/mskcc-omics-workflows/test-datasets/raw/neoantigen/neoantigen/Homo_sapiens.GRCh37.75.gtf.gz'
 }
