+++ modules/msk/generatemutfasta/1.2/meta.yml
@@ -6,11 +6,18 @@ keywords:
   - fasta
   - netmhc
   - mutation
+  - hgvsc
+  - Mutalyzer
 tools:
   - neoantigen_utils:
       description: "Collection of helper scripts for neoantigen processing"
       documentation: "https://github.com/mskcc-omics-workflows/modules"
       licence: [""]
+      identifier: ""
+  - mutalyzer:
+      description: "A tool primarily designed to check descriptions of sequence variants according to the Human Genome Sequence Variation Society (HGVS) guidelines."
+      documentation: "https://mutalyzer.readthedocs.io/en/latest/"
+      licence: ["MIT"]
       identifier: ""
 
 input:
@@ -22,15 +29,11 @@ input:
     - inputMaf:
         type: file
         description: Maf outputtted by Tempo that was run through phyloWGS
-        pattern: "*.{maf}"
-  - - cds:
+        pattern: "*.maf"
+  - - mutalyzer_cache:
         type: file
-        description: coding sequence resource fasta
-        pattern: "*.{cds.all.fa.gz}"
-    - cdna:
-        type: file
-        description: cDNA resource fasta
-        pattern: "*.{cdna.all.fa.gz}"
+        description: Mutatalyzer cache to use with the normalizer
+        pattern: "*.tar.gz"
 output:
   - mut_fasta:
       - meta:
@@ -41,7 +44,7 @@ output:
       - "*_out/*.MUT.sequences.fa":
           type: file
           description: Mutated fasta sequence
-          pattern: "*.{MUT.sequences.fa}"
+          pattern: "*.MUT.sequences.fa"
   - wt_fasta:
       - meta:
           type: map
@@ -51,7 +54,17 @@ output:
       - "*_out/*.WT.sequences.fa":
           type: file
           description: Wildtype fasta sequence
-          pattern: "*.{WT.sequences.fa}"
+          pattern: "*.WT.sequences.fa"
+  - mut_fasta_log:
+      - meta:
+          type: map
+          description: |
+            Groovy Map containing sample information
+            e.g. `[ id:'sample1', single_end:false ]`
+      - "*_out/*_generate_mut_fasta.log":
+          type: file
+          description: Log file for the mutated fasta generation
+          pattern: "*_generate_mut_fasta.log"
   - versions:
       - versions.yml:
           type: file
