+++ modules/msk/generatemutfasta/1.2/tests/nextflow.config
@@ -1,0 +1,5 @@
+process {
+    withName: 'GENERATEMUTFASTA' {
+        containerOptions = '--network none'
+    }
+}
