+++ modules/msk/generatemutfasta/1.2/tests/tags.yml
@@ -1,0 +1,2 @@
+generatemutfasta/1.2:
+  - "modules/msk/generatemutfasta/1.2/**"
