+++ modules/msk/mutalyzer/retriever/tests/tags.yml
@@ -1,0 +1,2 @@
+mutalyzer/retriever:
+  - "modules/msk/mutalyzer/retriever/**"
