--- modules/msk/neoantigenutils/generatemutfasta/tests/tags.yml
@@ -1,2 +1,0 @@
-neoantigenutils/generatemutfasta:
-  - "modules/msk/neoantigenutils/generatemutfasta/**"
