+++ subworkflows/msk/generate_mutated_peptides/tests/main.nf.test
@@ -1,0 +1,153 @@
+nextflow_workflow {
+
+    name "Test Subworkflow GENERATE_MUTATED_PEPTIDES"
+    script "../main.nf"
+    workflow "GENERATE_MUTATED_PEPTIDES"
+
+    tag "subworkflows"
+    tag "subworkflows_nfcore"
+    tag "subworkflows_msk"
+    tag "subworkflows/generate_mutated_peptides"
+    tag "mutalyzer/retriever"
+    tag "neoantigenutils/generatehlastring"
+    tag "generatemutfasta/1.2"
+    tag "neosv"
+
+
+    test("generate_mutated_peptides - maf,hla,fa,gff3,gtf - fa,str,log") {
+
+        when {
+            workflow {
+                """
+
+                input[0] = Channel.of([
+                    [ id:'test', single_end:false ], // meta map
+                    file(params.test_data_mskcc['neoantigen']['small_test_chr2and16']['somatic_filtered_maf'], checkIfExists: true),
+                    file(params.test_data_mskcc['neoantigen']['winners_hla_txt'], checkIfExists: true),
+                    ])
+
+                input[1] = Channel.value(
+                    file(params.test_data_mskcc['neoantigen']['small_test_chr2and16']['fa_gzip'], checkIfExists: true)
+                    )
+
+                input[2] = Channel.value(
+                    file(params.test_data_mskcc['neoantigen']['small_test_chr2and16']['gff3'], checkIfExists: true)
+                    )
+
+                input[3] = Channel.value(
+                    file(params.test_data_mskcc['neoantigen']['gtf'], checkIfExists: true)
+                    )
+
+                input[4] = Channel.value(
+                    file(params.test_data_mskcc['neoantigen']['cdna'], checkIfExists: true)
+                    )
+
+                """
+            }
+        }
+
+        then {
+            assertAll(
+                { assert workflow.success},
+                { assert snapshot(  workflow.out.mut_fasta[0][1],
+                                    workflow.out.wt_fasta[0][1],
+                                    file(workflow.out.mut_fasta_log[0][1]).name,
+                                    workflow.out.hla_string[0][1],
+                                ).match()
+                }
+            )
+        }
+    }
+
+    test("generate_mutated_peptides - maf,hla,sv,fa,gff3,gtf - fa,str,log") {
+
+        when {
+            workflow {
+                """
+                input[0] = Channel.of([
+                    [ id:'test', single_end:false ], // meta map
+                    file(params.test_data_mskcc['neoantigen']['small_test_chr2and16']['somatic_filtered_maf'], checkIfExists: true),
+                    file(params.test_data_mskcc['neoantigen']['winners_hla_txt'], checkIfExists: true),
+                    file(params.test_data_mskcc['neoantigen']['small_test_chr2and16']['sv_filtered_bedpe'], checkIfExists: true)
+                    ])
+
+                input[1] = Channel.value(
+                    file(params.test_data_mskcc['neoantigen']['small_test_chr2and16']['fa_gzip'], checkIfExists: true)
+                    )
+
+                input[2] = Channel.value(
+                    file(params.test_data_mskcc['neoantigen']['small_test_chr2and16']['gff3'], checkIfExists: true)
+                    )
+
+                input[3] = Channel.value(
+                    file(params.test_data_mskcc['neoantigen']['gtf'], checkIfExists: true)
+                    )
+
+                input[4] = Channel.value(
+                    file(params.test_data_mskcc['neoantigen']['cdna'], checkIfExists: true)
+                    )
+
+                """
+            }
+        }
+
+        then {
+            assertAll(
+                { assert workflow.success},
+                { assert snapshot(  workflow.out.mut_fasta[0][1],
+                                    workflow.out.wt_fasta[0][1],
+                                    file(workflow.out.mut_fasta_log[0][1]).name,
+                                    workflow.out.sv_mut_fasta[0][1],
+                                    workflow.out.sv_wt_fasta[0][1],
+                                    workflow.out.hla_string[0][1],
+                                ).match()
+                }
+            )
+        }
+    }
+
+    test("generate_mutated_peptides - maf,hla,sv,fa,gff3,gtf - fa,str,log - stub") {
+
+        options "-stub"
+
+        when {
+            workflow {
+                """
+
+                input[0] = Channel.of([
+                    [ id:'test', single_end:false ], // meta map
+                    file('somatic_filtered_maf', checkIfExists: false),
+                    file('winners_hla_txt', checkIfExists: false),
+                    file('sv_filtered_bedpe', checkIfExists: false)
+                    ])
+
+                input[1] = Channel.value(
+                    file('fa_gzip', checkIfExists: false)
+                    )
+
+                input[2] = Channel.value(
+                    file('gff3', checkIfExists: false)
+                    )
+
+                input[3] = Channel.value(
+                    file('gtf', checkIfExists: false)
+                    )
+
+                input[4] = Channel.value(
+                    file('cdna', checkIfExists: false)
+                    )
+
+                """
+            }
+        }
+
+        then {
+            assertAll(
+                { assert workflow.success},
+                { assert snapshot(workflow.out).match()}
+            )
+        }
+
+
+    }
+}
